[wallet-refactor-5] Remove Mod and unify WalletModal (#2061)
* Remove MOD for WalletModal

* Remove WalletModalMod

* Fix code style issues with Prettier

---------
